Avanço na implementação das tabelas e rotas.

+++ api/routes/disciplina.py
@@ -1,16 +1,21 @@
 from http import HTTPStatus
 
 from fastapi import APIRouter, HTTPException
-from sqlalchemy import select
+from sqlalchemy import func, select
 from sqlalchemy.exc import IntegrityError
 
 from api.deps import CurrentUser, GetSession
+from core.utils import update_schema
+from models.concurso import Concurso
+from models.concurso_disciplina import ConcursoDisciplina
 from models.disciplina import Disciplina
 from schemas.disciplina import (
     DisciplinaCreate,
     DisciplinaList,
     DisciplinaPublic,
+    DisciplinaUpdate,
 )
+from schemas.utils import Message
 
 router = APIRouter(prefix='/disciplinas', tags=['disciplinas'])
 
@@ -27,7 +32,7 @@ async def create_disciplina(
     db_disciplina = await session.scalar(
         select(Disciplina).where(
             (Disciplina.nome == disciplina.nome)
-            & (Disciplina.id_user_created == current_user.id)
+            & (Disciplina.usuario_id == current_user.id)
         )
     )
 
@@ -40,7 +45,7 @@ async def create_disciplina(
     try:
         db_disciplina = Disciplina(
             nome=disciplina.nome,
-            id_user_created=current_user.id,  # Usa o ID do usuário autenticado
+            usuario_id=current_user.id,  # Usa o ID do usuário autenticado
         )
 
         session.add(db_disciplina)
@@ -59,7 +64,7 @@ async def create_disciplina(
 async def read_disciplinas(session: GetSession, current_user: CurrentUser):
     # Retorna apenas as disciplinas do usuário autenticado
     result = await session.execute(
-        select(Disciplina).where(Disciplina.id_user_created == current_user.id)
+        select(Disciplina).where(Disciplina.usuario_id == current_user.id)
     )
     disciplinas = result.scalars().all()
     return {'disciplinas': disciplinas}
@@ -72,7 +77,7 @@ async def read_disciplina(
     db_disciplina = await session.scalar(
         select(Disciplina).where(
             (Disciplina.id == disciplina_id)
-            & (Disciplina.id_user_created == current_user.id)
+            & (Disciplina.usuario_id == current_user.id)
         )
     )
 
@@ -88,14 +93,14 @@ async def read_disciplina(
 @router.put('/{disciplina_id}', response_model=DisciplinaPublic)
 async def update_disciplina(
     disciplina_id: int,
-    disciplina_update: DisciplinaCreate,
+    disciplina_update: DisciplinaUpdate,
     session: GetSession,
     current_user: CurrentUser,
 ):
     db_disciplina = await session.scalar(
         select(Disciplina).where(
             (Disciplina.id == disciplina_id)
-            & (Disciplina.id_user_created == current_user.id)
+            & (Disciplina.usuario_id == current_user.id)
         )
     )
 
@@ -109,7 +114,7 @@ async def update_disciplina(
     existing_disciplina = await session.scalar(
         select(Disciplina).where(
             (Disciplina.nome == disciplina_update.nome)
-            & (Disciplina.id_user_created == current_user.id)
+            & (Disciplina.usuario_id == current_user.id)
             & (Disciplina.id != disciplina_id)
         )
     )
@@ -120,27 +125,30 @@ async def update_disciplina(
             detail='Você já possui outra disciplina com este nome.',
         )
 
-    db_disciplina.nome = disciplina_update.nome
+    db_disciplina = update_schema(
+        schema=disciplina_update, model=db_disciplina
+    )
 
     try:
         await session.commit()
         await session.refresh(db_disciplina)
         return db_disciplina
-    except IntegrityError:
+    except IntegrityError as e:
+        print(e)
         raise HTTPException(
             status_code=HTTPStatus.BAD_REQUEST,
             detail='Erro ao atualizar disciplina.',
         )
 
 
-@router.delete('/{disciplina_id}', status_code=HTTPStatus.NO_CONTENT)
+@router.delete('/{disciplina_id}', response_model=Message)
 async def delete_disciplina(
     disciplina_id: int, session: GetSession, current_user: CurrentUser
 ):
     db_disciplina = await session.scalar(
         select(Disciplina).where(
             (Disciplina.id == disciplina_id)
-            & (Disciplina.id_user_created == current_user.id)
+            & (Disciplina.usuario_id == current_user.id)
         )
     )
 
@@ -150,7 +158,19 @@ async def delete_disciplina(
             detail='Disciplina não encontrada.',
         )
 
+    vinculado = await session.scalar(
+        select(ConcursoDisciplina)
+        .where(ConcursoDisciplina.disciplina_id == disciplina_id)
+        .limit(1)
+    )
+
+    if vinculado:
+        raise HTTPException(
+            status_code=HTTPStatus.CONFLICT,
+            detail='Não é possível alterar disciplina vinculada a um concurso.'
+        )
+
     await session.delete(db_disciplina)
     await session.commit()
 
-    return None
+    return Message(message='Disciplina deletada com sucesso.')
